Move admin state into module
